# openapi/agh.json
{
...
    "info": {
        "title": "AGH API",
        "version": "1.0.0"
    },
    "openapi": "3.0.3",
    "paths": {
...
        "/api/extensions": {
            "get": {
                "operationId": "listExtensions",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "extensions": {
                                            "items": {
                                                "properties": {
                                                    "actions": {
                                                        "items": {
                                                            "type": "string"
                                                        },
                                                        "type": "array"
                                                    },
                                                    "capabilities": {
                                                        "items": {
                                                            "type": "string"
                                                        },
                                                        "type": "array"
                                                    },
                                                    "daemon_running": {
                                                        "type": "boolean"
                                                    },
                                                    "enabled": {
                                                        "type": "boolean"
                                                    },
                                                    "health": {
                                                        "type": "string"
                                                    },
                                                    "health_message": {
                                                        "type": "string"
                                                    },
                                                    "last_error": {
                                                        "type": "string"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "pid": {
                                                        "type": "integer"
                                                    },
                                                    "source": {
                                                        "type": "string"
                                                    },
                                                    "state": {
                                                        "type": "string"
                                                    },
                                                    "type": {
                                                        "type": "string"
                                                    },
                                                    "uptime_seconds": {
                                                        "format": "int64",
                                                        "type": "integer"
                                                    },
                                                    "version": {
                                                        "type": "string"
                                                    }
                                                },
                                                "required": [
                                                    "daemon_running",
                                                    "enabled",
                                                    "name",
                                                    "source",
                                                    "state",
                                                    "type",
                                                    "version"
                                                ],
                                                "type": "object"
                                            },
                                            "type": "array"
                                        }
                                    },
                                    "required": ["extensions"],
                                    "type": "object"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "required": ["error"],
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Internal server error"
                    },
                    "503": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "required": ["error"],
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Extension service is not configured"
                    },
                    "default": {
                        "description": ""
                    }
                },
                "summary": "List installed extensions",
                "tags": ["extensions"],
                "x-agh-transports": ["uds"]
            },
            "post": {
                "operationId": "installExtension",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "checksum": {
                                        "type": "string"
                                    },
                                    "path": {
                                        "type": "string"
                                    }
                                },
                                "required": ["checksum", "path"],
                                "type": "object"
                            }
                        }
                    },
                    "description": "JSON request body",
                    "required": true
                },
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "extension": {
                                            "properties": {
                                                "actions": {
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "type": "array"
                                                },
                                                "capabilities": {
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "type": "array"
                                                },
                                                "daemon_running": {
                                                    "type": "boolean"
                                                },
                                                "enabled": {
                                                    "type": "boolean"
                                                },
                                                "health": {
                                                    "type": "string"
                                                },
                                                "health_message": {
                                                    "type": "string"
                                                },
                                                "last_error": {
                                                    "type": "string"
                                                },
                                                "name": {
                                                    "type": "string"
                                                },
                                                "pid": {
                                                    "type": "integer"
                                                },
                                                "source": {
                                                    "type": "string"
                                                },
                                                "state": {
                                                    "type": "string"
                                                },
                                                "type": {
                                                    "type": "string"
                                                },
                                                "uptime_seconds": {
                                                    "format": "int64",
                                                    "type": "integer"
                                                },
                                                "version": {
                                                    "type": "string"
                                                }
                                            },
                                            "required": [
                                                "daemon_running",
                                                "enabled",
                                                "name",
                                                "source",
                                                "state",
                                                "type",
                                                "version"
                                            ],
                                            "type": "object"
                                        }
                                    },
                                    "required": ["extension"],
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Created"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "required": ["error"],
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Invalid install request"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "required": ["error"],
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Internal server error"
                    },
                    "503": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "required": ["error"],
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Extension service is not configured"
                    },
                    "default": {
                        "description": ""
                    }
                },
                "summary": "Install an extension by path and checksum",
                "tags": ["extensions"],
                "x-agh-transports": ["uds"]
            }
        },
...
        "/api/extensions/{name}/disable": {
            "post": {
                "operationId": "disableExtension",
                "parameters": [
                    {
                        "description": "Extension name",
                        "in": "path",
                        "name": "name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "extension": {
                                            "properties": {
                                                "actions": {
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "type": "array"
                                                },
                                                "capabilities": {
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "type": "array"
                                                },
                                                "daemon_running": {
                                                    "type": "boolean"
                                                },
                                                "enabled": {
                                                    "type": "boolean"
                                                },
                                                "health": {
                                                    "type": "string"
                                                },
                                                "health_message": {
                                                    "type": "string"
                                                },
                                                "last_error": {
                                                    "type": "string"
                                                },
                                                "name": {
                                                    "type": "string"
                                                },
                                                "pid": {
                                                    "type": "integer"
                                                },
                                                "source": {
                                                    "type": "string"
                                                },
                                                "state": {
                                                    "type": "string"
                                                },
                                                "type": {
                                                    "type": "string"
                                                },
                                                "uptime_seconds": {
                                                    "format": "int64",
                                                    "type": "integer"
                                                },
                                                "version": {
                                                    "type": "string"
                                                }
                                            },
                                            "required": [
                                                "daemon_running",
                                                "enabled",
                                                "name",
                                                "source",
                                                "state",
                                                "type",
                                                "version"
                                            ],
                                            "type": "object"
                                        }
                                    },
                                    "required": ["extension"],
                                    "type": "object"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "required": ["error"],
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Extension not found"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "required": ["error"],
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Internal server error"
                    },
                    "503": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "required": ["error"],
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Extension service is not configured"
                    },
                    "default": {
                        "description": ""
                    }
                },
                "summary": "Disable an installed extension",
                "tags": ["extensions"],
                "x-agh-transports": ["uds"]
            }
        },
...
    },
...
}